moving project to root
