+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "core-concepts-app",
+  "name": "global-react-js",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
